Calculate neighbors and leave plug for thermals
The thermal model still needs to be done, but I left a placeholder
function for now. Maybe later we will have it so you can swap out
thermal models, but that is more work than we really need to do right
now. It is more important to have it log metrics, such as average
agent temperature over time, because that would be good to have for the
report.