Add environment and framework (#13)
* Ignore logging-fstring-interpolation linting

Fstrings are the preferred string formatting method in Python, so I'm
not sure why this is linted.

* Add conda environment

* Ignore too-few-public-methods

It is annoying and I don't care about it for this project.

* Change to integer positions

* Change move back to float

I forgot this was direction and velocity, not positions. Maybe it should
be a target position instead? That could be interesting.

* Add environment setup

Still need to test. Not sure if this will work... But it should be
enough to get the idea of how it will work across.

* Add gif saving and misc small features

* Calculate neighbors and leave plug for thermals

The thermal model still needs to be done, but I left a placeholder
function for now. Maybe later we will have it so you can swap out
thermal models, but that is more work than we really need to do right
now. It is more important to have it log metrics, such as average
agent temperature over time, because that would be good to have for the
report.

* Add environment thermal parameters

* Add thermal parameters for agents

* Fix naming of files from config name