Implement the movement policy for our final design (#19)
Policy detail see #17.

Updated `Agent.get_move()` by implementing the movement policy
by Sid's priliminary simulation. Detail see issue comment.

Added `Environment.update_thermal()` to test the movement policy.
This is yet to be completed because it should not be this simple.

Changed `cfg/template.ini` to test the movement policy. This
set-up should be good for the swarming cluster to be seen pretty
fast.

Co-authored-by: Wayne Stegner <51062260+stegnerw@users.noreply.github.com>